+++ Phonebook/Service/WorkerService.cs
@@ -1,7 +1,9 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using Phonebook.Data;
 using Phonebook.Domain.Services;
+using Phonebook.Exceptions;
 using Phonebook.Models;
+using System.Net;
 
 namespace Phonebook.Service
 {
@@ -15,17 +17,32 @@ namespace Phonebook.Service
 
         public async Task<Worker> Create(Worker worker)
         {
+            if(worker == null)
+            {
+                throw new HttpResponseException(HttpStatusCode.BadRequest, "Worker is null");
+            }
             await _context.Workers.AddAsync(worker);
             await _context.SaveChangesAsync();
             return worker;
         }
 
-        public async Task<Worker> Delete(int id)
+        public async Task<bool> Delete(int id)
         {
+
+            string query = "SELECT * FROM Director WHERE Id = {0}";
+            var department = await _context.Departments
+                .FromSqlRaw(query, id)
+                .AsNoTracking()
+                .FirstAsync();
+            if (department == null)
+            {
+                throw new HttpResponseException(HttpStatusCode.NotFound, "Director is not found");
+            }
+
             var workerToDelete = await _context.Workers.FindAsync(id);
             _context.Workers.Remove(workerToDelete);
             await _context.SaveChangesAsync();
-            return workerToDelete;
+            return true;
         }
 
         public async Task<IEnumerable<Worker>> Get()
@@ -40,16 +57,29 @@ namespace Phonebook.Service
                 .FromSqlRaw(query, departmentId)
                 .AsNoTracking()
                 .ToListAsync();
+            if (workers.Count == 0)
+            {
+                throw new HttpResponseException(HttpStatusCode.NotFound, "Workers is not found");
+            }
             return workers;
         }
 
         public async Task<Worker> GetById(int id)
         {
+            if (await _context.Workers.FindAsync(id) == null)
+            {
+                throw new HttpResponseException(HttpStatusCode.NotFound, "Worker is not found");
+            }
             return await _context.Workers.FindAsync(id);
         }
 
-        public async Task<Worker> Update(Worker worker)
+        public async Task<Worker> Update(int id, Worker worker)
         {
+
+            if (await _context.Workers.FindAsync(id) == null)
+            {
+                throw new HttpResponseException(HttpStatusCode.NotFound, "Worker is not found");
+            }
             _context.Entry(worker).State = EntityState.Modified;
             await _context.SaveChangesAsync();
             return _context.Workers.FirstOrDefault();
